… spdy/conn_test.go
… spdy/copy.go