This change include use of a new msgpack implementation based on a fork of github.com/vmihailenco/msgpack.  The changes were to support libchan were extended types, interface encoding, raw messages, and removing of a non-standard time encoding.

Additional cleanup changes to support this PR include removing of the raw connection encoding in the spdy implementation which was never defined in the libchan protocol.

Includes:
- Different msgpack implementation
- New raw message type
- New extension implementation

What this enables:
- More efficient proxying
- Zero copy in-memory channel transfer
- Partial Decoding
- Sending/receiving non struct/map objects

fixes #47