@dmcgowan 2015-04-04
In memory pipe benchmark results
~~~
benchmark              old ns/op     new ns/op     delta
BenchmarkRoundTrip     21897         1868          -91.47%

benchmark              old allocs     new allocs     delta
BenchmarkRoundTrip     20             0              -100.00%

benchmark              old bytes     new bytes     delta
BenchmarkRoundTrip     752           0             -100.00%
~~~